# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: AI Career Guidance API for pre-university students in Zimbabwe
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /insights:
    get:
      operationId: getCareerInsights
      tags: [jobs]
      summary: Get job market insights for a career
      parameters:
        - name: career
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job market insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerInsights"
  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: Get job listings for a career
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobListing"
...
components:
  schemas:
...
    CareerInsights:
      type: object
      properties:
        career:
          type: string
        demandLevel:
          type: string
          enum: [High, Medium, Low]
        jobCount:
          type: integer
        topSkills:
          type: array
          items:
            type: string
        averageSalary:
          type: ["string", "null"]
      required: [career, demandLevel, jobCount, topSkills]
    JobListing:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        location:
          type: string
        salary:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
          description: Link to the job advert
      required: [title, company, location]
...